multimodal_model/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==24.3.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
datasets==3.2.0
deprecation==2.1.0
dill==0.3.8
durationpy==0.9
einops==0.8.0
fastapi==0.115.7
filelock==3.17.0
flatbuffers==25.1.21
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
google-auth==2.38.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.3
keras==3.8.0
kserve==0.14.1
kubernetes==32.0.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.1.3
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.2.0
prometheus_client==0.20.0
propcache==0.2.1
protobuf==4.25.5
psutil==5.9.8
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
sentencepiece==0.2.0
setuptools==70.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
timing-asgi==0.3.1
timm==1.0.14
tokenizers==0.21.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
transformers==4.48.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
